requirements.dev.txt
tornado==5.1.1
pyeva==3.0.1
sqlalchemy==1.2.12
sqlalchemy-utils==0.33.5
psycopg2-binary==2.7.5
bravado_core==5.0.7
nose2==0.8.0
pylint==2.1.1
flake8==3.5.0
coverage==4.5.1
requests
